requirements.txt
azure-cosmos>=4.2.0